requirements.txt
# Core ML and Data Science
tensorflow>=2.10.0
keras>=2.10.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.5.0

# Image Processing
opencv-python>=4.6.0
Pillow>=9.0.0
pydicom>=2.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Progress bars and utilities
tqdm>=4.64.0

# GUI Framework
PyQt5>=5.15.0

# Report Generation
reportlab>=3.6.0
jinja2>=3.1.0

# Additional utilities
scipy>=1.9.0
albumentations>=1.3.0

# Note: Removed imgaug as it conflicts with albumentations
# Both provide similar functionality, albumentations is more stable
